scripts/train/train.py
#!/usr/bin/env python
"""
IndexTTS2 训练脚本
"""

import argparse
import torch
import yaml
from pathlib import Path


def parse_args():
    parser = argparse.ArgumentParser(description='IndexTTS2 训练脚本')
    parser.add_argument('--config', type=str, required=True,
                        help='配置文件路径')
    parser.add_argument('--gpu', type=str, default='0',
                        help='使用的GPU ID，多个用逗号分隔')
    parser.add_argument('--resume', type=str, default=None,
                        help='恢复训练的检查点路径')
    parser.add_argument('--distributed', action='store_true',
                        help='是否使用分布式训练')
    return parser.parse_args()


def load_config(config_path):
    """加载配置文件"""
    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    return config


def main():
    args = parse_args()
    
    # 加载配置
    config = load_config(args.config)
    
    # 设置设备
    if torch.cuda.is_available() and config['device']['use_gpu']:
        device = torch.device('cuda')
        print(f"使用GPU: {args.gpu}")
    else:
        device = torch.device('cpu')
        print("使用CPU")
    
    # TODO: 加载模型
    # model = IndexTTS2(config['model'])
    # model = model.to(device)
    
    # TODO: 加载数据
    # train_loader = ...
    # val_loader = ...
    
    # TODO: 设置优化器
    # optimizer = ...
    
    # TODO: 训练循环
    # for epoch in range(config['training']['epochs']):
    #     train_one_epoch(...)
    #     validate(...)
    
    print("训练脚本模板已创建，请根据实际实现填充代码")


if __name__ == '__main__':
    main()
