requirements.txt
# IndexTTS2 项目依赖

# 深度学习框架
torch>=1.10.0
torchaudio>=0.10.0
numpy>=1.21.0
scipy>=1.7.0

# 音频处理
librosa>=0.9.0
soundfile>=0.10.0
pyaudio>=0.2.11

# 数据处理
pandas>=1.3.0
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.4.0
tensorboard>=2.7.0
seaborn>=0.11.0

# 工具库
tqdm>=4.62.0
pyyaml>=5.4.0
omegaconf>=2.1.0
hydra-core>=1.1.0

# 文本处理
jieba>=0.42.1  # 中文分词（如需要）
pypinyin>=0.44.0  # 中文拼音（如需要）

# 其他工具
pillow>=8.3.0
imageio>=2.9.0

# 开发和测试
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
